+++ src/CryptoNoteCore/Blockchain.cpp
@@ -236,7 +236,7 @@ public:
         s(twap_lo, "twap_lo");
         s(twap_hi, "twap_hi");
         if (s.type() == ISerializer::INPUT)
-          m_bs.m_twapAccumulator = ((unsigned __int128)twap_hi << 64) | twap_lo;
+          m_bs.m_twapAccumulator = ((uint128_t)twap_hi << 64) | twap_lo;
       }
       s(m_bs.m_piState, "pi_state");
       s(m_bs.m_heatLaunchTwap, "heat_launch_twap");
@@ -3283,7 +3283,7 @@ bool Blockchain::pushBlock(const Block &blockData, const std::vector<Transaction
   // TWAP accumulation per block (v11+) — spot price stored in Q64.64
   if (block.bl.majorVersion >= BLOCK_MAJOR_VERSION_11 && !m_ammPool.isEmpty()) {
     uint64_t spotPrice = ammGetSpotPrice(m_ammPool.reserveXfg, m_ammPool.reserveHeat);
-    unsigned __int128 q64 = ((unsigned __int128)spotPrice << 64) / 1000000000000000000ULL;
+    uint128_t q64 = ((uint128_t)spotPrice << 64) / 1000000000000000000ULL;
     m_twapAccumulator += q64;
     m_twapBlockCount++;
   }
@@ -3336,7 +3336,7 @@ bool Blockchain::pushBlock(const Block &blockData, const std::vector<Transaction
 
       computeNewRedemptionPrice(m_piState, marketPrice, targetRatio, epochDuration, parameters::HEAT_STABILITY_MODE);
     }
-    int128_t epochTwapAvg = (m_twapBlockCount > 0) ? (int128_t)(m_twapAccumulator / m_twapBlockCount) : 0;
+    int128_t epochTwapAvg = (m_twapBlockCount > 0) ? (int128_t)(m_twapAccumulator / m_twapBlockCount) : int128_t(uint64_t(0));
     // Reset TWAP for next epoch
     m_twapAccumulator = 0;
     m_twapBlockCount = 0;
@@ -3989,11 +3989,11 @@ bool Blockchain::pushBlock(BlockEntry &block) {
     }
 
     // Compute fee rates for this epoch on both tracks.
-    // Use __uint128_t for the intermediate product to prevent uint64_t overflow
+    // Use uint128_t for the intermediate product to prevent uint64_t overflow
     uint64_t epochFeeRate = 0;
     if (epochCdLocked > 0 && regularCdShare > 0) {
       epochFeeRate = static_cast<uint64_t>(
-          (__uint128_t)regularCdShare * CryptoNote::parameters::FEE_POOL_RATE_PRECISION / epochCdLocked);
+          (uint128_t)regularCdShare * CryptoNote::parameters::FEE_POOL_RATE_PRECISION / epochCdLocked);
     }
     m_commitmentIndex.recordEpochFeeRate(epochNumber, epochFeeRate, regularCdShare, epochCdLocked);
 
@@ -4001,7 +4001,7 @@ bool Blockchain::pushBlock(BlockEntry &block) {
     uint64_t legacyEpochFeeRate = 0;
     if (epochLegacyBondLocked > 0 && legacyBondShare > 0) {
       legacyEpochFeeRate = static_cast<uint64_t>(
-          (__uint128_t)legacyBondShare * CryptoNote::parameters::FEE_POOL_RATE_PRECISION / epochLegacyBondLocked);
+          (uint128_t)legacyBondShare * CryptoNote::parameters::FEE_POOL_RATE_PRECISION / epochLegacyBondLocked);
     }
     m_commitmentIndex.recordLegacyEpochFeeRate(epochNumber, legacyEpochFeeRate, legacyBondShare, epochLegacyBondLocked);
 
@@ -4072,9 +4072,9 @@ bool Blockchain::pushBlock(BlockEntry &block) {
       if (expectedRatioScaled == 0) goto arb_done;
 
       for (int round = 0; round < 40; ++round) {
-        uint64_t poolRatioScaled = (m_ammPool.reserveXfg * 10000) / std::max(m_ammPool.reserveHeat, 1ULL);
+        uint64_t poolRatioScaled = (m_ammPool.reserveXfg * 10000) / std::max(m_ammPool.reserveHeat, uint64_t(1));
         int64_t  deviation       = ((int64_t)poolRatioScaled - (int64_t)expectedRatioScaled) * 10000
-                                   / std::max(expectedRatioScaled, 1ULL);
+                                   / std::max(expectedRatioScaled, uint64_t(1));
         if (deviation < 500 && deviation > -500) break;  // within 0.05%, stop
 
         // Arb size: 3% of smaller reserve, max 5% of XFG reserve
@@ -4298,7 +4298,7 @@ void Blockchain::popBlock(const Crypto::Hash& blockHash) {
     m_protocolLpShares = snap.protocolLpShares;
     m_treasuryLpYield = snap.treasuryLpYield;
     m_bootstrapRepaymentVault = snap.bootstrapRepaymentVault;
-    m_twapAccumulator = ((unsigned __int128)snap.twapAccumulatorHi << 64) | snap.twapAccumulatorLo;
+    m_twapAccumulator = ((uint128_t)snap.twapAccumulatorHi << 64) | snap.twapAccumulatorLo;
     m_twapBlockCount = snap.twapBlockCount;
     m_ammPool.reserveXfg = snap.ammReserveXfg;
     m_ammPool.reserveHeat = snap.ammReserveHeat;
@@ -4700,8 +4700,8 @@ void Blockchain::popTransaction(const Transaction& transaction, const Crypto::Ha
         if (swap.direction == 0) {
           uint64_t preXfg = m_ammPool.reserveXfg - swap.inputAmount;
           uint64_t feeAdj = parameters::HEARTH_FEE_DIVISOR - feeBps;
-          uint64_t outputAmount = (uint64_t)(((unsigned __int128)m_ammPool.reserveHeat
-            * swap.inputAmount * feeAdj) / ((unsigned __int128)preXfg * parameters::HEARTH_FEE_DIVISOR));
+          uint64_t outputAmount = (uint64_t)(((uint128_t)m_ammPool.reserveHeat
+            * swap.inputAmount * feeAdj) / ((uint128_t)preXfg * parameters::HEARTH_FEE_DIVISOR));
           uint64_t outputNoFee = ammGetOutputAmount(swap.inputAmount, preXfg, m_ammPool.reserveHeat, 0);
           uint64_t actualFee = outputNoFee > outputAmount ? outputNoFee - outputAmount : 0;
           if (m_ammPool.accumulatedLpFees >= actualFee) m_ammPool.accumulatedLpFees -= actualFee;
@@ -4710,8 +4710,8 @@ void Blockchain::popTransaction(const Transaction& transaction, const Crypto::Ha
         } else {
           uint64_t preHeat = m_ammPool.reserveHeat - swap.inputAmount;
           uint64_t feeAdj = parameters::HEARTH_FEE_DIVISOR - feeBps;
-          uint64_t outputAmount = (uint64_t)(((unsigned __int128)m_ammPool.reserveXfg
-            * swap.inputAmount * feeAdj) / ((unsigned __int128)preHeat * parameters::HEARTH_FEE_DIVISOR));
+          uint64_t outputAmount = (uint64_t)(((uint128_t)m_ammPool.reserveXfg
+            * swap.inputAmount * feeAdj) / ((uint128_t)preHeat * parameters::HEARTH_FEE_DIVISOR));
           uint64_t outputNoFee = ammGetOutputAmount(swap.inputAmount, preHeat, m_ammPool.reserveXfg, 0);
           uint64_t actualFee = outputNoFee > outputAmount ? outputNoFee - outputAmount : 0;
           if (m_ammPool.accumulatedLpFees >= actualFee) m_ammPool.accumulatedLpFees -= actualFee;
